Move BuildsRoute.query to PackageController+Queries
