+++ Sources/App/Controllers/PackageController.swift
@@ -106,84 +106,6 @@ struct PackageController {
 }
 
 
-// TODO: move
-extension PackageController {
-    enum BuildsRoute {
-        struct PackageInfo: Equatable {
-            var packageName: String?
-            var repositoryOwner: String
-            var repositoryName: String
-
-            static func query(on database: Database, owner: String, repository: String) -> EventLoopFuture<(PackageInfo)> {
-                Joined3<Package, Repository, Version>
-                    .query(on: database, version: .defaultBranch)
-                    .filter(Repository.self, \.$owner, .custom("ilike"), owner)
-                    .filter(Repository.self, \.$name, .custom("ilike"), repository)
-                    .first()
-                    .unwrap(or: Abort(.notFound))
-                    .flatMapThrowing { model in
-                        let repo = model.repository
-                        guard let repoOwner = repo.owner,
-                              let repoName = repo.name else {
-                                  throw Abort(.notFound)
-                              }
-                        return .init(packageName: model.version.packageName,
-                                     repositoryOwner: repoOwner,
-                                     repositoryName: repoName)
-                    }
-            }
-        }
-
-        struct BuildInfo: Equatable {
-            var versionKind: Version.Kind
-            var reference: Reference
-            var buildId: Build.Id
-            var swiftVersion: SwiftVersion
-            var platform: Build.Platform
-            var status: Build.Status
-
-            static func query(on database: Database, owner: String, repository: String) -> EventLoopFuture<[BuildInfo]> {
-                Joined4<Build, Version, Package, Repository>
-                    .query(on: database)
-                    .filter(Version.self, \Version.$latest != nil)
-                    .filter(Repository.self, \.$owner, .custom("ilike"), owner)
-                    .filter(Repository.self, \.$name, .custom("ilike"), repository)
-                    .field(\.$id)
-                    .field(\.$swiftVersion)
-                    .field(\.$platform)
-                    .field(\.$status)
-                    .field(Version.self, \.$latest)
-                    .field(Version.self, \.$packageName)
-                    .field(Version.self, \.$reference)
-                    .all()
-                    .flatMapThrowing { results in
-                        try results
-                            .compactMap { res -> BuildInfo? in
-                                let build = res.build
-                                let version = res.version
-                                guard let kind = version.latest,
-                                      let reference = version.reference else {
-                                          return nil
-                                      }
-                                return try BuildInfo(versionKind: kind,
-                                                     reference: reference,
-                                                     buildId: build.requireID(),
-                                                     swiftVersion: build.swiftVersion,
-                                                     platform: build.platform,
-                                                     status: build.status)
-                            }
-                    }
-            }
-        }
-
-        static func query(on database: Database, owner: String, repository: String) -> EventLoopFuture<(PackageInfo, [BuildInfo])> {
-            PackageInfo.query(on: database, owner: owner, repository: repository)
-                .and(BuildInfo.query(on: database, owner: owner, repository: repository))
-        }
-    }
-}
-
-
 // MARK: - PackageResult
